UI Tweaks - position admin, search bar, timesheet tables.
- Moved position delete into edit dialog (positions are never deleted.)
- incorporated position form component directly into the position template.
- Fixed searching person by id so the person is shown in the results, and not just blanks.
- Swapped from 100% table width to auto on the person manage timesheet & missing timesheet tables.
- Added text labels to the edit & trash icons on the timesheet tables.
- Fixed spelling bug on the person manage header.
