+++ app/templates/admin/positions.hbs
@@ -1,50 +1,49 @@
-
 <h1>Clubhouse Positions</h1>
 
-{{#if position}}
-  <PositionForm @position={{position}} @trainingPositions={{trainingPositions}} @onSave={{action "save"}} @onCancel={{action "cancel"}} />
-{{/if}}
-
-<div class="row mb-2">
-  <div class="col-sm-3">
-    <ChForm::Select @name="typeFilter" @value={{typeFilter}} @options={{typeOptions}} @onChange={{action (mut typeFilter)}} @controlClass="form-control" />
+<div class="row mb-3">
+  <label class="col-form-label col-auto">Type filter</label>
+  <div class="col-auto">
+    <ChForm::Select @name="typeFilter" @value={{typeFilter}} @options={{typeOptions}} @onChange={{action
+            (mut typeFilter)}} @controlClass="form-control"/>
   </div>
-  <div class="col-sm-9 text-right">
-    <button type="button" class="btn btn-primary" {{action "new"}}>New Position</button>
+  <div class="col-auto">
+    <button type="button" class="btn btn-primary" {{on "click"  this.newAction}}>New Position</button>
   </div>
 </div>
 
-<p class="text-muted">Showing {{viewPositions.length}} of {{positions.length}}</p>
-<table class="table table-hover table-sm">
+<table class="table table-striped table-hover table-sm">
+  <caption>Showing {{this.viewPositions.length}} of {{this.positions.length}} positions</caption>
   <thead>
-    <tr>
-      <th>Title</th>
-      <th>Short Title</th>
-      <th>Type</th>
-      <th>Min / Max</th>
-      <th>New User Eligible</th>
-      <th>All Rangers</th>
-      <th>Count Hours</th>
-      <th>Auto Signout</th>
-      <th>On SL report</th>
-      <th>Prevent Multi. Enrollments</th>
-      <th>Training / Full Email</th>
-      <th>Action</th>
-    </tr>
+  <tr>
+    <th>ID</th>
+    <th>Title</th>
+    <th>Short Title</th>
+    <th>Type</th>
+    <th>Min / Max</th>
+    <th class="text-center">New User Eligible</th>
+    <th class="text-center">All Rangers</th>
+    <th class="text-center">Count Hrs</th>
+    <th class="text-center">Auto Signout</th>
+    <th class="text-center">On SL report</th>
+    <th>Prevent Multi. Enrollments</th>
+    <th>Training / Contact Email</th>
+    <th>Action</th>
+  </tr>
   </thead>
 
   <tbody>
-    {{#each viewPositions as |position|}}
-    <tr class="position-line">
+  {{#each viewPositions as |position|}}
+    <tr>
+      <td class="text-right">{{position.id}}</td>
       <td>{{position.title}}</td>
-      <td>{{position.short_title}}</td>
+      <td>{{present-or-not position.short_title "-"}}</td>
       <td>{{position.type}}</td>
-      <td class="text-right">{{position.min}} / {{position.max}}</td>
-      <td>{{yesno position.new_user_eligible}}</td>
-      <td>{{yesno position.all_rangers}}</td>
-      <td>{{yesno position.count_hours}}</td>
-      <td>{{yesno position.auto_signout}}</td>
-      <td>{{yesno position.on_sl_report}}</td>
+      <td class="text-center">{{position.min}} / {{position.max}}</td>
+      <td class="text-center">{{yesno position.new_user_eligible}}</td>
+      <td class="text-center">{{yesno position.all_rangers}}</td>
+      <td class="text-center">{{yesno position.count_hours}}</td>
+      <td class="text-center">{{yesno position.auto_signout}}</td>
+      <td class="text-center">{{yesno position.on_sl_report}}</td>
       <td>{{if position.prevent_multiple_enrollments "Y" "-"}}</td>
       <td>
         {{#if position.training_position_id}}
@@ -55,10 +54,79 @@
         {{/if}}
       </td>
       <td class="actions">
-        <button class="btn btn-primary" {{action "edit" position}} title="Edit position">{{fa-icon "edit"}}</button>
-        <button class="btn btn-danger" {{action "delete" position}} title="Delete position">{{fa-icon "trash-alt"}}</button>
+        <button type="button" class="btn btn-primary btn-sm" {{on "click" (fn this.editAction position)}}
+                title="Edit position">{{fa-icon "edit"}} Edit
+        </button>
       </td>
     </tr>
-    {{/each}}
+  {{/each}}
   </tbody>
 </table>
+
+{{#if this.position}}
+  <ModalDialog @title={{if this.position.isNew "Create Position" (concat "Edit Position ID #" this.position.id " - " this.position.title)}} as |Modal|>
+    <ChForm @formId="position-form" @originalModel={{this.position}}
+            @validator={{this.positionValidations}} @onSubmit={{this.saveAction}}
+            @onCancel={{this.cancelAction}} as |f|>
+      <Modal.body>
+        <fieldset>
+          <legend>Description</legend>
+          <div class="form-row">
+            <f.input @name="title" @label="Title" @maxlength=40 @size=40 @grid="col-auto"
+                     @hint={{concat (if f.model.title f.model.title.length "0") " of 40 characters."}} />
+            <f.input @name="short_title" @label="Short Title" @maxlength=6 @grid="col-auto"
+                     @hint={{concat (if f.model.short_title f.model.title.short_title "0") " of 6 characters."}} />
+            <f.input @name="type" @type="select" @label="Type" @options={{this.positionTypes}} @grid="col-auto"/>
+          </div>
+        </fieldset>
+
+        <fieldset>
+          <legend>Suggested Ranger Counts Per Slot</legend>
+          <div class="form-row">
+            <f.input @name="min" @label="Min" @size=3 @maxlength=3 @grid="col-auto"/>
+            <f.input @name="max" @label="Max" @size=3 @maxlength=3 @grid="col-auto"/>
+          </div>
+        </fieldset>
+
+        <fieldset>
+          <legend>Flags</legend>
+          <div class="form-row mb-2">
+            <f.input @name="all_rangers" @type="checkbox" @label="All rangers" @inline=true/>
+            <f.input @name="new_user_eligible" @type="checkbox" @label="New user eligible" @inline=true/>
+            <f.input @name="count_hours" @type="checkbox" @label="Count hours" @inline=true/>
+            <f.input @name="auto_signout" @type="checkbox" @label="Can be auto-signed out" @inline=true/>
+            <f.input @name="on_sl_report" @type="checkbox" @label="On SL report" @inline=true/>
+            <f.input @name="prevent_multiple_enrollments" @type="checkbox"
+                     @label="Prevent Multiple Enrollments"
+                     @inline=true/>
+          </div>
+        </fieldset>
+
+        <fieldset>
+          <legend>Training</legend>
+          <p class="form-row">
+            The following section is relevant when type is "Training";
+          </p>
+          <div class="form-row">
+            <f.input @name=training_position_id" @label="Training Position" @type="select"
+                     @options={{this.trainingOptions}}
+                     @grid="col-auto"/>
+            <f.input @name="slot_full_email" @label="Shift Full Email" @type="text" @size=40 @maxlength=200
+                     @grid="col-auto"/>
+          </div>
+        </fieldset>
+      </Modal.body>
+
+      <Modal.footer @noAlign=true>
+        <f.submit @label={{if position.isNew "Create" "Update"}} />
+        <f.cancel/>
+        <div class="ml-auto">
+          <button type="button" class="btn btn-danger" {{on "click" this.deleteAction}}>{{fa-icon "trash"}} Delete
+            Position
+          </button>
+        </div>
+      </Modal.footer>
+    </ChForm>
+  </ModalDialog>
+{{/if}}
+
