--- app/templates/components/position-form.hbs
@@ -1,51 +1,0 @@
-{{#ch-form "position" position modalBox=true modalTitle=formTitle validator=positionValidations onSubmit=(action "save") onCancel=(action "cancel") as |f|}}
-  <fieldset>
-    <legend>Description</legend>
-    <div class="form-row">
-      {{f.input "title" label="Title" maxlength=40 size=40 grid="col-auto"
-        hint=(concat (if f.model.title f.model.title.length "0") " of 40 characters.")
-      }}
-      {{f.input "short_title" label="Short Title" maxlength=6 grid="col-auto"
-        hint=(concat (if f.model.short_title f.model.title.short_title "0") " of 6 characters.")
-      }}
-      {{f.input "type" type="select" label="Type" options=positionTypes grid="col-auto"}}
-    </div>
-  </fieldset>
-
-  <fieldset>
-    <legend>Suggested Ranger Counts Per Slot</legend>
-    <div class="form-row">
-      {{f.input "min" label="Min" size=3 maxlength=3 grid="col-auto"}}
-      {{f.input "max" label="Max" size=3 maxlength=3 grid="col-auto"}}
-    </div>
-  </fieldset>
-
-  <fieldset>
-    <legend>Flags</legend>
-    <div class="form-row mb-2">
-      {{f.input "all_rangers" type="checkbox" label="All rangers" inline=true}}
-      {{f.input "new_user_eligible" type="checkbox" label="New user eligible" inline=true}}
-      {{f.input "count_hours" type="checkbox" label="Count hours" inline=true}}
-      {{f.input "auto_signout" type="checkbox" label="Can be auto-signed out" inline=true}}
-      {{f.input "ol_sl_report" type="checkbox" label="On SL report" inline=true}}
-    </div>
-  </fieldset>
-
-  <fieldset>
-    <legend>Training</legend>
-    <p class="form-row">
-      The following section is relevant when type is "Training";
-    </p>
-    <div class="form-row">
-      {{f.input "training_position_id" label="Training Position" type="select" options=trainingOptions grid="col-auto"}}
-      {{f.input "slot_full_email" label="Shift Full Email" type="text" size=40 maxlength=200 grid="col-auto"}}
-      {{f.input "prevent_multiple_enrollments" type="checkbox" label="Prevent Multiple Enrollments" inline=true}}
-    </div>
-  </fieldset>
-
-  <div class="mt-4">
-    <f.submit @label={{if position.isNew "Create" "Update"}} />
-    {{f.cancel}}
-  </div>
-
-{{/ch-form}}
