+++ app/templates/person.hbs
@@ -32,7 +32,7 @@
       {{#if this.person.user_authorized}}
         has the status "{{person.status}}"
       {{else}}
-        is disable
+        is disabled
       {{/if}}
       and is not accessible. Contact the Personnel Manager: {{mail-to (config "PersonnelEmail")}}.
     </ChAlert>
