--- tests/integration/components/position-form-test.js
@@ -1,26 +1,0 @@
-import { module, skip /* test */ } from 'qunit';
-import { setupRenderingTest } from 'ember-qunit';
-import { render } from '@ember/test-helpers';
-import hbs from 'htmlbars-inline-precompile';
-
-module('Integration | Component | position-form', function(hooks) {
-  setupRenderingTest(hooks);
-
-  skip('it renders', async function(assert) {
-    // Set any properties with this.set('myProperty', 'value');
-    // Handle any actions with this.set('myAction', function(val) { ... });
-
-    await render(hbs`{{position-form}}`);
-
-    assert.dom(this.element).hasText('');
-
-    // Template block usage:
-    await render(hbs`
-      {{#position-form}}
-        template block text
-      {{/position-form}}
-    `);
-
-    assert.dom(this.element).hasText('template block text');
-  });
-});
